remove mesa-asahi-edge
Since mesa 25.1, support for asahi is enabled by default. This means,
the nixpkgs mesa is sufficient and we don't need any custom mesa, or
methods to replace it.

In addition to the now ineffective
hardware.asahi.useExperimentalGPUDriver option, this also removes the
hardware.asahi.experimentalGPUInstallMode option, which was already
deprecated before.